Modify : 추천 받을 때, 별점 저장 수정  - eatda_be/vo/Group.java : 그룹별 groupId 및 reviewId 생성위한 group 테이블  - eatda_be/vo/MaxStoreId.java : reviewId용 maxValue 추가  - eatda_be/vo/Review.java : review 업데이트용 테이블 추가  - eatda_be/vo/Score.java : reviewId 추가  - eatda_be/vo/Store.java : float -> double  - eatda_be/repository/GroupRepository.java : groupId로 찾기 추가  - eatda_be/repository/ReviewUpdateRepository.java : review테이블 repository  - eatda_be/service/MeetingServiceImpl.java : 추천에 groupId설정 추가  - eatda_be/service/ReviewServiceImpl.java : review 및 store 업데이트 추가  - eatda_be/service/UserServiceImpl.java : 회원가입시 review 테이블에 reviewId에 맞는 튜플 추가

+++ eatda_be/src/main/java/com/ssafy/eatda/vo/Store.java
@@ -8,11 +8,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Store {
-    private String storeId;
-    private String storeName;
-    private String storeAddress;
-    private float storeLatitude;
-    private float storeLongitude;
-    private float avgRate;
-    private int reviewCount;
+  private String storeId;
+  private String storeName;
+  private String storeAddress;
+  private double storeLatitude;
+  private double storeLongitude;
+  private float avgRate;
+  private int reviewCount;
 }
